requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
cssbeautifier==1.15.4
distro==1.9.0
Django==5.2.3
django-browser-reload==1.18.0
django-extensions==4.1
django-livereload-server==0.5.1
django-widget-tweaks==1.5.0
djlint==1.36.4
EditorConfig==0.17.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsbeautifier==1.15.4
json5==0.12.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
openai==1.93.0
pathspec==0.12.1
pillow==11.2.1
pydantic==2.11.7
pydantic_core==2.33.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
